% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charlson Comorbidity Index.R
\name{cci}
\alias{cci}
\title{cci()}
\usage{
cci(data, comorbidity, age)
}
\arguments{
\item{data}{Your data file in which Charlson Comorbidity index is to be calculated}

\item{comorbidity}{A vector of all comorbidity variables}

\item{age}{The name of the age variable}
}
\value{
data: a new data.frame named "data". This data frame contains two new variables: "cci_1987" & "cci_2011"
}
\description{
This file aims to calculate the Charlson Comorbidity index(1985 orginal version and the 2011 Quan version)
}
\details{
To calculate the Charlson Comorbidity index ("cci_1987" 1987 orginal version and cci_2011 the 2011 Quan version)
"cci_1987": The Charlson Comorbidity index, developed by Mary E. Charlson in 1987.
"cci_2011": The Charlson Comorbidity index, updated by Hude Quan in 2011.
}
\author{
Miao Cai <email: miao.cai@slu.edu>
}
