% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eci.R
\name{eci}
\alias{eci}
\title{To calculate the Elixhauser Comorbidity Index}
\usage{
eci(data, comorbidity)
}
\arguments{
\item{data}{Your data file in which Elixhauser Comorbidity Index is to be calculated}

\item{comorbidity}{A vector of all comorbidity variables}
}
\value{
data: a new data.frame named "data". This data frame contains a new variable "Elix_Index": The Elixhauser Comorbidity Index, developed by Anne Elixhauser in 1998
}
\description{
This file aims to calculate the Elixhauser Comorbidity Index.
}
\note{
The Elixhauser Comorbidity Index does not include age as a component of the index.
}
\references{
Elixhauser, A., Steiner, C., Harris, D. R., & Coffey, R. M. (1998). Comorbidity measures for use with administrative data. Medical care, 36(1), 8-27.

van Walraven, C., Austin, P. C., Jennings, A., Quan, H., & Forster, A. J. (2009). A modification of the Elixhauser comorbidity measures into a point system for hospital death using administrative data. Medical care, 626-633.

Quan, H., Sundararajan, V., Halfon, P., Fong, A., Burnand, B., Luthi, J. C., ... & Ghali, W. A. (2005). Coding algorithms for defining comorbidities in ICD-9-CM and ICD-10 administrative data. Medical care, 1130-1139.
}
\author{
Miao Cai <email: miao.cai@slu.edu>
}
