% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3.R
\name{c3}
\alias{c3}
\title{To calculate the c3 Comobidity Index (targeted at cancer patients)}
\usage{
c3(data, comorbidity)
}
\arguments{
\item{data}{Your data file in which Elixhauser Comorbidity Index is to be calculated}

\item{comorbidity}{A vector of all comorbidity variables}
}
\value{
data: a new data.frame named "data". This data frame contains a new variable "c3": the c3 Index was developed by Sarfati et al. in 2014. This is a comorbidity index targeted at cancer patients.
}
\description{
This file aims to calculate the c3 Comobidity Index.
}
\note{
This C3 index is used specifically for cancer patients.
}
\references{
Sarfati, D., et al., Cancer-specific administrative data-based comorbidity indices provided valid alternative to Charlson and National Cancer Institute Indices. J Clin Epidemiol, 2014. 67(5): p. 586-95

Sarfati, D., Developing new comorbidity indices for cancer populations using administrative data. 2013, University of Otago: Dunedin. <Dr. Sarfati's doctoral dissertation>
}
\author{
Miao Cai <email: miao.cai@outlook.com>
}
