% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCI.R
\name{cci}
\alias{cci}
\title{To calculate the Charlson Comorbidity index(1985 orginal version and the 2011 Quan version)}
\usage{
cci(data_file, x, age)
}
\arguments{
\item{data_file}{Your data file in which Charlson Comorbidity index is to be calculated}

\item{x}{A vector of all comorbidity variables}

\item{age}{The name of the age variable}
}
\value{
cci_1987: The Charlson Comorbidity index, developed by Mary E. Charlson in 1987

cci_2011: The Charlson Comorbidity index, updated by Hude Quan in 2011
}
\description{
This file aims to calculate the Charlson Comorbidity index(1985 orginal version and the 2011 Quan version)
}
\author{
Miao Cai <email: miaocai@slu.edu>
}

